Move TestAuthServer to helpers/
