+++ tests/auth.test.ts
@@ -3,7 +3,7 @@ import TestResolver from '../src/system/TestResolver'
 import { Config } from '../src/utils/config'
 import authCommand from '../src/commands/auth'
 import { authEndpoint, graphcoolConfigFilePath, systemAPIEndpoint } from '../src/utils/constants'
-import TestAuthServer from '../src/api/TestAuthServer'
+import TestAuthServer from './helpers/test_auth_server'
 import { testToken } from './fixtures/mock_data'
 import { TestSystemEnvironment } from '../src/types'
 import TestOut from '../src/system/TestOut'
