+++ tests/helpers/test_auth_server.ts
@@ -1,5 +1,5 @@
-import {AuthServer} from '../types'
-import {testToken} from '../../tests/mock_data/mockData'
+import {AuthServer} from '../../src/types'
+import {testToken} from '../fixtures/mock_data'
 
 export default class TestAuthServer implements AuthServer {
